+++ platformio/platforms/titiva.py
@@ -16,12 +16,14 @@ class TitivaPlatform(BasePlatform):
 
         "toolchain-gccarmnoneeabi": {
             "path": join("tools", "toolchain"),
+            "alias": "toolchain",
             "default": True
         },
 
         "tool-lm4flash": {
             "path": join("tools", "lm4flash"),
-            "default": True,
+            "alias": "uploader",
+            "default": True
         },
 
         "framework-energiativa": {